Change the behavior when a measure is selected
When a measure is selected, instead of using it as a frequency, we will instead use it as the value directly instead of the category's value. Therefore, instead of computing a cumulative percentile over time, we will now compute the percentiles of the range of values for that period of time.

As an example, suppose we have a dataset where we have X number of activities per day. Previously, we would show a chart with a cumulative distribution over time which allowed you to see relative rates of change over the time period. The 50th percentile in that case would be the date at which the sum of all X values from the beginning of the range was half the value of the sum of the X values across the entire range. Now, we instead show the percentiles of the distibution of those X values per day such that the 50th percentile shows the median value of X across all days. This should align better with the standard use cases.